learn-rights-python/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
pydantic>=2.0
pymongo>=4.0
pyjwt>=2.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7
google-generativeai>=0.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
google-auth>=2.0.0
streamlit>=1.28.0
requests>=2.28.0
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99
protobuf>=3.20.0
